requirements.txt
azure.functions
larksuite-oapi
flask
PyYAML
openai